refactor(hal): change hal structure
