+++ hal/.cargo/config.toml
@@ -1,7 +1,3 @@
 [build]
 target = "riscv64gc-unknown-none-elf"
 
-[target.riscv64gc-unknown-none-elf]
-rustflags = [
-    "-Clink-arg=-Tsrc/linker.ld", "-Cforce-frame-pointers=yes"
-]
