+++ hal/build.rs
@@ -1,0 +1,3 @@
+fn main() {
+    
+}
