+++ hal/src/arch/riscv64/mod.rs
@@ -1,0 +1,4 @@
+pub mod mem;
+pub mod instruction;
+
+pub use crate::hal::instruction::Instruction;
