+++ hal/src/hal/mod.rs
@@ -1,0 +1,3 @@
+pub mod mem;
+pub mod trap;
+pub mod instruction;
