+++ hal/src/hal/trap.rs
