+++ os/src/arch/riscv64/mod.rs
@@ -1,7 +1,1 @@
-use crate::mm::VirtAddr;
-
 pub mod entry;
-pub mod mem;
-pub mod instruction;
-
-pub use instruction::Instruction;
